Refactor: 코드 리팩토링
1. cart.js navbar 추가
2. changeUserInfo.js 로고이미지 추가
3. slideShow.js 기능 수정
4. manageCategory.js 확인창 추가, 코드 리팩토링
5. newProductPage.js 이미지 경로 수정 및 콘솔로그 삭제
6. product & product-detail.js 이미지 경로 수정

+++ src/views/manageCategory/manageCategory.js
@@ -53,12 +53,14 @@ async function categoryLanding() {
   editButtons.forEach((el) =>
     el.addEventListener("click", async (e) => {
       editModal.classList.add("is-active");
-      const prevCategory = e.path[2].children[1].innerText;
+      const prevCategory = e.target.classList[0];
+      editCategoryName.value = prevCategory;
 
       // 카테고리 수정 이벤트 리스너
       editBtn.addEventListener("click", async () => {
         const data = {
-          foodType: editCategoryName.value,
+          // 음식 카테고리의 종류가 같다면 카테고리의 설명만 변경하도록 foodType 값을 null로 지정
+          foodType: editCategoryName.value === prevCategory ? null : editCategoryName.value,
           description: editCategoryDescription.value,
         };
         try {
@@ -74,9 +76,12 @@ async function categoryLanding() {
   // 클릭 시 카테고리 삭제 api 요청 후 카테고리 삭제
   deleteButtons.forEach((el) =>
     el.addEventListener("click", async (e) => {
-      const catecory = e.path[2].children[1].innerText;
-      await Api.delete("/api/category/" + catecory);
-      window.location.href = "/admin/manageCategory/";
+      const catecory = e.target.classList[0];
+      if (confirm("카테고리를 삭제하시겠습니까?")) {
+        await Api.delete("/api/category/" + catecory);
+        alert("카테고리가 삭제되었습니다.");
+        window.location.href = "/admin/manageCategory/";
+      }
     })
   );
 }
@@ -88,7 +93,8 @@ async function addCategoryFn() {
     description: addCategoryDescription.value,
   };
   try {
-    await Api.post("/api/category/add", data);
+    await Api.post("/api/category/", data);
+    alert("카테고리가 추가되었습니다.");
     window.location.href = "/admin/manageCategory/";
   } catch (err) {
     alert(err);
@@ -104,10 +110,10 @@ function createCategoryList(data) {
     <div class="column is-2 order-summary">${el.foodType}</div>
     <div class="column is-5">${el.description}</div>
     <div class="column is-1">
-      <button class="button editButton">수정</button>
+      <button class="${el.foodType} button editButton">수정</button>
     </div>
     <div class="column is-1">
-      <button class="button deleteButton">삭제</button>
+      <button class="${el.foodType} button deleteButton">삭제</button>
     </div>
   </div>
   `
